migrated silence command to history added silence to FastAPI
