+++ api_docs/post_command_silence_host.md
@@ -1,0 +1,9 @@
+__/api/command/silence_host/<host_id>/<minutes>__ - sets the given hosts silenced property to True for the given amount of minutes. This will silence any notifications for this time.
+
+```
+{
+  "is_silenced": true,
+  "success": true,
+  "until": "05-02-2023 01:31PM"
+}
+```
